RunMultipleTimes.pl being fixed.
git-svn-id: http://seqsee1.googlecode.com/svn/Seqsee/branches/feature_extend_long_relations@741 19b06628-3642-0410-8fdd-33ae7300e995

+++ util/RunMultipleTimes.pl
@@ -5,6 +5,8 @@ use Test::Seqsee;
 
 my ( $times, $terms ) = @ARGV;
 
+my $TIMES_TO_RUN = 10;
+
 my $MW = new MainWindow();
 
 my $Text = $MW->Text( -width => 130 )->pack();
@@ -12,41 +14,65 @@ $Text->tagConfigure( 'success',   -background => 'blue',   -foreground => 'white
 $Text->tagConfigure( 'error',     -background => 'red',    -foreground => 'white' );
 $Text->tagConfigure( 'nosuccess', -background => 'yellow', -foreground => 'white' );
 
-$Text->insert('end', "Please wait....");
+$Text->insert( 'end', "Please wait...." );
 $Text->update();
+my @RESULTS;
 
 StartRun();
 
 $MW->MainLoop();
 
-sub StartRun {
-    my ( $imp_ref, $worse_ref, $results_ref, $opts_ref )
-        = RegHarness( { seq => $terms, max_steps => 20000 } );
+use Storable;
 
-    my $seq = join( ', ', @{ $opts_ref->{seq} } );
-    ### results: $results_ref
-    ### opts_ref: $opts_ref
-    ### seq: $seq
-    Update($results_ref);
+sub StartRun {
+    my ( $seq, $continuation ) = split( /\|/, $terms );
+    my @cmd = (
+        'c:\perl\bin\perl',   'util/RunTestOnce.pl',
+        qq{--seq="$seq"},     qq{--continuation="$continuation"},
+        qq{-max_steps=10000}, qq{--min_extension=3},
+        qq{--max_false=3}
+    );
+
+    # my $cmd = join(" ", @cmd);
+    for ( 1 .. $TIMES_TO_RUN ) {
+        print ">>@cmd<<\n";
+        system @cmd;
+        open( my $RESULT, '<', "foo" ) or confess "Unable to open file >>foo<<";
+        my $result_str        = join('', <$RESULT>);
+        my $result_object = Storable::thaw($result_str) or confess "Unable to thaw: >>$result_str<<!!";
+        push @RESULTS, $result_object;
+        Update();
+    }
 }
 
 sub Update {
-    my $results_ref = shift;
-    $Text->delete( '0.0', 'end' );
-    Display_Selected( $results_ref );
-    $MW->update();
+    if (@RESULTS) {
+        $Text->delete( '0.0', 'end' );
+        my @top_row = map { find_tag( $_->get_status() ) } @RESULTS;
+        $Text->insert( 'end', @top_row, scalar(@RESULTS), '', "/$TIMES_TO_RUN", '', "\n" );
+        for (@RESULTS) {
+            $Text->insert( 'end', "\n............\n", '',  find_tag( $_->get_status() ),
+                "\t", '', $_->get_steps(), '', "\n\n" );
+            $Text->insert('end', $_->get_error()) if $_->get_error();
+        }
+        $MW->update();
+    }
 }
 
 sub find_tag {
-    my ($msg) = @_;
-    if ( $msg =~ /^SUCCESS/ ) {
-        return ["success"];
+    my ($status) = @_;
+    print $status;
+    if ( $status->IsSuccess() ) {
+        return ( "O", ["success"] );
     }
-    elsif ( $msg =~ /^Error/ ) {
-        return ["error"];
+    elsif ( $status->IsACrash() ) {
+        return ( " CRASH! ", ["error"] );
+    }
+    elsif ( $status->IsAtLeastAnExtension() ) {
+        return ( "?", ["success"] );
     }
     else {
-        return ["nosuccess"];
+        return ( "X", ["nosuccess"] );
     }
 }
 
@@ -54,13 +80,13 @@ sub Display_Selected {
     my ($arr_ref) = @_;
     my @arr = @$arr_ref;
 
-    my @string = map { ( "X", find_tag($_) ) } @arr;
-    $Text->insert( 'end', @string, "\n" );
+    my @string = map { ( " X ", find_tag($_) ) } @arr;
+    $Text->insert( 'end', @string, " \n " );
 
     for my $line (@arr) {
         ### line: $line
-        $Text->insert( 'end', "======\n\n" );
-        $Text->insert( 'end', $line, '', "\n" );
+        $Text->insert( 'end', " == == == \n \n " );
+        $Text->insert( 'end', $line, '', " \n " );
     }
 }
